+++ atomic/atom/startupinit/src/startupinit.js
@@ -37,7 +37,9 @@ module.exports = async (...args) => {
           let err;
           let connection = {};
           for (let [kconconf, conconf] of Object.entries(setting.db)) {
-            for (let [key, dbconf] of Object.entries(conconf)) {
+            for (let [key, dbconf, restore = false] of Object.entries(
+              conconf
+            )) {
               dbconf["engine"] = "sqlite3";
               dbconf["setting"] = setting;
               dbconf["dbgroup"] = kconconf;
@@ -49,8 +51,12 @@ module.exports = async (...args) => {
                 if (conn.code != 0) err += conn.message;
                 else {
                   connection = { ...connection, ...conn.data };
-                  if (!(await conn.data[key].ischema()))
-                    conn.data[key].import(join(pathname, "data", `${key}.sql`));
+                  if (restore) {
+                    if (!(await conn.data[key].ischema()))
+                      conn.data[key].import(
+                        join(pathname, "data", `${key}.sql`)
+                      );
+                  }
                 }
               }
             }
@@ -69,7 +75,7 @@ module.exports = async (...args) => {
       };
 
       lib["mariadb"] = async (...args) => {
-        const [[setting, compname], pathname, dbengine] = args;
+        const [[setting, compname, restore = false], pathname, dbengine] = args;
         let output = handler.dataformat;
         try {
           let err;
@@ -87,12 +93,14 @@ module.exports = async (...args) => {
                 if (conn.code != 0) err += conn.message;
                 else {
                   connection = { ...connection, ...conn.data };
-                  if (!(await conn.data[key].ischema(key))) {
-                    await conn.data[key].import(
-                      join(pathname, "data", `${key}.sql`)
-                    );
-                    if (!(await conn.data[key].ischema(key)))
-                      err += "Import database failure!";
+                  if (restore) {
+                    if (!(await conn.data[key].ischema(key))) {
+                      await conn.data[key].import(
+                        join(pathname, "data", `${key}.sql`)
+                      );
+                      if (!(await conn.data[key].ischema(key)))
+                        err += "Import database failure!";
+                    }
                   }
                   connection[key].disconnect();
                 }