+++ Changelog.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.5] - 2025-11-26
+
+### Summary
+
+- Focus to workflow engine design.
+
+### Added
+
+- Client and Server Side Event to atomic/atom complete design. Completed on 2025-11-27
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+- Bug fix in compmgr engine. Completed on 2025-11-26
+
+### Security
+
+[1.1.5]: https://github.com/wkloh76/oricommjs_v2/releases/tag/1.1.5
+
 ## [1.1.4] - 2025-11-14
 
 ### Summary